refactor: preference => priority
Also delete endpoint in rest.js
